+++ source/wiz/__main__.py
@@ -1,13 +1,11 @@
 # :coding: utf-8
 
-import sys
-
 import wiz.command_line
 
 
 def main():
     """Execute main command line interface passing command line arguments."""
-    wiz.command_line.main(sys.argv[1:])
+    wiz.command_line.main(prog_name="wiz")
 
 
 if __name__ == "__main__":